Refactor p2
