+++ randomizer/Patching/CosmeticColors.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import gzip
 import random
-import zlib
-from random import randint
 from typing import TYPE_CHECKING, List, Tuple
 from io import BytesIO
 
@@ -13,11 +11,9 @@ from PIL import Image, ImageDraw, ImageEnhance
 
 import js
 from randomizer.Enums.Kongs import Kongs
-from randomizer.Enums.Settings import CharacterColors, ColorblindMode, RandomModels, KongModels, WinConditionComplex
-from randomizer.Enums.Models import Model, Sprite
+from randomizer.Enums.Settings import CharacterColors, ColorblindMode, KongModels, WinConditionComplex
 from randomizer.Enums.Maps import Maps
 from randomizer.Enums.Types import BarrierItems
-from randomizer.Patching.generate_kong_color_images import convertColors
 from randomizer.Patching.Cosmetics.CustomTextures import writeTransition, writeCustomPaintings, writeCustomPortal
 from randomizer.Patching.Cosmetics.Krusha import placeKrushaHead, fixBaboonBlasts, kong_index_mapping, fixModelSmallKongCollision
 from randomizer.Patching.Cosmetics.Colorblind import (
@@ -36,47 +32,28 @@ from randomizer.Patching.Cosmetics.Colorblind import (
     recolorRotatingRoomTiles,
 )
 from randomizer.Patching.Lib import (
-    PaletteFillType,
-    SpawnerChange,
-    applyCharacterSpawnerChanges,
     compatible_background_textures,
     TableNames,
-    getRawFile,
 )
 from randomizer.Patching.LibImage import (
     getImageFile, 
-    TextureFormat, 
-    getRandomHueShift, 
-    hueShift, 
+    TextureFormat,
     ExtraTextures, 
     getBonusSkinOffset, 
     writeColorImageToROM, 
-    numberToImage, 
-    getRGBFromHash,
+    numberToImage,
     maskImage,
     maskImageWithColor,
     getKongItemColor,
-    hueShiftImageContainer,
+    hueShiftColor,
 )
 from randomizer.Patching.Patcher import ROM, LocalROM
 from randomizer.Settings import Settings
 from randomizer.Patching.Cosmetics.ModelSwaps import (
-    turtle_models,
-    panic_models,
-    bother_models,
-    piano_models,
-    piano_extreme_model,
-    spotlight_fish_models,
-    candy_cutscene_models,
-    funky_cutscene_models,
-    funky_cutscene_models_extreme,
-    boot_cutscene_models,
-    melon_random_sprites,
+    model_mapping,
+    applyCosmeticModelSwaps,
 )
-
-if TYPE_CHECKING:
-    from PIL.Image import Image
-
+from randomizer.Patching.Cosmetics.KongColor import writeKongColors, changeModelTextures
 
 class HelmDoorSetting:
     """Class to store information regarding helm doors."""
@@ -109,112 +86,7 @@ class HelmDoorImages:
         self.dimensions = dimensions
         self.format = format
 
-
-
-
-model_mapping = {
-    KongModels.default: 0,
-    KongModels.disco_chunky: 6,
-    KongModels.krusha: 7,
-    KongModels.krool_cutscene: 9,
-    KongModels.krool_fight: 8,
-    KongModels.cranky: 10,
-    KongModels.candy: 11,
-    KongModels.funky: 12,
-}
-krusha_texture_replacement = {
-    # Textures Krusha can use when he replaces various kongs (Main color, belt color)
-    Kongs.donkey: (3724, 0x177D),
-    Kongs.diddy: (4971, 4966),
-    Kongs.lanky: (3689, 0xE9A),
-    Kongs.tiny: (6014, 0xE68),
-    Kongs.chunky: (3687, 3778),
-}
-model_texture_sections = {
-    KongModels.krusha: {
-        "skin": [0x4738, 0x2E96, 0x3A5E],
-        "kong": [0x3126, 0x354E, 0x37FE, 0x41E6],
-    },
-    KongModels.krool_fight: {
-        "skin": [
-            0x61D6,
-            0x63FE,
-            0x6786,
-            0x7DD6,
-            0x7E8E,
-            0x7F3E,
-            0x7FEE,
-            0x5626,
-            0x56E6,
-            0x5A86,
-            0x5BAE,
-            0x5D46,
-            0x5E2E,
-            0x5FAE,
-            0x69BE,
-            0x735E,
-            0x7C5E,
-            0x7E4E,
-            0x7EF6,
-            0x7FA6,
-            0x8056,
-        ],
-        "kong": [0x607E, 0x7446, 0x7D46, 0x80FE],
-    },
-    # KongModels.krool_cutscene: {
-    #     "skin": [0x4A6E, 0x4CBE, 0x52AE, 0x55BE, 0x567E, 0x57E6, 0x5946, 0x5AA6, 0x5E06, 0x5EC6, 0x6020, 0x618E, 0x62F6, 0x6946, 0x6A6E, 0x6C5E, 0x6D86, 0x6F76, 0x702E, 0x70DE, 0x718E, 0x72FE, 0x4FBE, 0x51FE, 0x5C26, 0x6476, 0x6826, 0x6B26, 0x6E3E, 0x6FE6, 0x7096, 0x7146, 0x71F6, 0x733E, 0x743E],
-    #     "kong": [],
-    # }
-}
-
-
-class KongPalette:
-    """Class to store information regarding a kong palette."""
-
-    def __init__(self, name: str, image: int, fill_type: PaletteFillType, alt_name: str = None):
-        """Initialize with given parameters."""
-        self.name = name
-        self.image = image
-        self.fill_type = fill_type
-        self.alt_name = alt_name
-        if alt_name is None:
-            self.alt_name = name
-
-
-class KongPaletteSetting:
-    """Class to store information regarding the kong palette setting."""
-
-    def __init__(self, kong: str, kong_index: int, palettes: list[KongPalette]):
-        """Initialize with given parameters."""
-        self.kong = kong
-        self.kong_index = kong_index
-        self.palettes = palettes.copy()
-        self.setting_kong = kong
-
-
-def getKongColor(settings: Settings, index: int):
-    """Get color index for a kong."""
-    kong_colors = ["#ffd700", "#ff0000", "#1699ff", "#B045ff", "#41ff25"]
-    mode = settings.colorblind_mode
-    if mode != ColorblindMode.off and settings.override_cosmetics:
-        if mode == ColorblindMode.prot:
-            kong_colors = ["#000000", "#0072FF", "#766D5A", "#FFFFFF", "#FDE400"]
-        elif mode == ColorblindMode.deut:
-            kong_colors = ["#000000", "#318DFF", "#7F6D59", "#FFFFFF", "#E3A900"]
-        elif mode == ColorblindMode.trit:
-            kong_colors = ["#000000", "#C72020", "#13C4D8", "#FFFFFF", "#FFA4A4"]
-    return kong_colors[index]
-
-
-DEFAULT_COLOR = "#000000"
-KLAPTRAPS = [Model.KlaptrapGreen, Model.KlaptrapPurple, Model.KlaptrapRed]
 RECOLOR_MEDAL_RIM = False
-
-
-def getRandomKlaptrapModel() -> Model:
-    """Get random klaptrap model."""
-    return random.choice(KLAPTRAPS)
-
 
 def changePatchFace(settings: Settings):
     """Change the top of the dirt patch image."""
@@ -238,103 +110,14 @@ def changePatchFace(settings: Settings):
 
 def apply_cosmetic_colors(settings: Settings):
     """Apply cosmetic skins to kongs."""
-    bother_model_index = Model.KlaptrapGreen
-    panic_fairy_model_index = Model.BananaFairy
-    panic_klap_model_index = Model.KlaptrapGreen
-    turtle_model_index = Model.Turtle
-    sseek_klap_model_index = Model.KlaptrapGreen
-    fungi_tomato_model_index = Model.Tomato
-    caves_tomato_model_index = Model.IceTomato
-    racer_beetle = Model.Beetle
-    racer_rabbit = Model.Rabbit
-    piano_burper = Model.KoshKremlingRed
-    spotlight_fish_model_index = Model.SpotlightFish
-    candy_model_index = Model.Candy
-    funky_model_index = Model.Funky
-    boot_model_index = Model.Boot
-    melon_sprite = Sprite.BouncingMelon
-    swap_bitfield = 0
-
+    
     ROM_COPY = ROM()
     sav = settings.rom_data
 
+    applyCosmeticModelSwaps(settings, ROM_COPY)
     changePatchFace(settings)
-
-    model_inverse_mapping = {}
-    for model in model_mapping:
-        val = model_mapping[model]
-        model_inverse_mapping[val] = model
-    ROM_COPY.seek(settings.rom_data + 0x1B8)
-    settings.kong_model_dk = model_inverse_mapping[int.from_bytes(ROM_COPY.readBytes(1), "big")]
-    settings.kong_model_diddy = model_inverse_mapping[int.from_bytes(ROM_COPY.readBytes(1), "big")]
-    settings.kong_model_lanky = model_inverse_mapping[int.from_bytes(ROM_COPY.readBytes(1), "big")]
-    settings.kong_model_tiny = model_inverse_mapping[int.from_bytes(ROM_COPY.readBytes(1), "big")]
-    settings.kong_model_chunky = model_inverse_mapping[int.from_bytes(ROM_COPY.readBytes(1), "big")]
-    if settings.override_cosmetics:
-        model_setting = RandomModels[js.document.getElementById("random_models").value]
-    else:
-        model_setting = settings.random_models
-    if model_setting == RandomModels.random:
-        bother_model_index = getRandomKlaptrapModel()
-    elif model_setting == RandomModels.extreme:
-        bother_model_index = random.choice(bother_models)
-        racer_beetle = random.choice([Model.Beetle, Model.Rabbit])
-        racer_rabbit = random.choice([Model.Beetle, Model.Rabbit])
-        if racer_rabbit == Model.Beetle:
-            spawner_changes = []
-            # Fungi
-            rabbit_race_fungi_change = SpawnerChange(Maps.FungiForest, 2)
-            rabbit_race_fungi_change.new_scale = 50
-            rabbit_race_fungi_change.new_speed_0 = 70
-            rabbit_race_fungi_change.new_speed_1 = 136
-            spawner_changes.append(rabbit_race_fungi_change)
-            # Caves
-            rabbit_caves_change = SpawnerChange(Maps.CavesChunkyIgloo, 1)
-            rabbit_caves_change.new_scale = 40
-            spawner_changes.append(rabbit_caves_change)
-            applyCharacterSpawnerChanges(spawner_changes)
-    if model_setting != RandomModels.off:
-        panic_fairy_model_index = random.choice(panic_models)
-        turtle_model_index = random.choice(turtle_models)
-        panic_klap_model_index = getRandomKlaptrapModel()
-        sseek_klap_model_index = getRandomKlaptrapModel()
-        fungi_tomato_model_index = random.choice([Model.Tomato, Model.IceTomato])
-        caves_tomato_model_index = random.choice([Model.Tomato, Model.IceTomato])
-        referenced_piano_models = piano_models.copy()
-        referenced_funky_models = funky_cutscene_models.copy()
-        if model_setting == RandomModels.extreme:
-            referenced_piano_models.extend(piano_extreme_model)
-            spotlight_fish_model_index = random.choice(spotlight_fish_models)
-            referenced_funky_models.extend(funky_cutscene_models_extreme)
-            boot_model_index = random.choice(boot_cutscene_models)
-        piano_burper = random.choice(referenced_piano_models)
-        candy_model_index = random.choice(candy_cutscene_models)
-        funky_model_index = random.choice(funky_cutscene_models)
-    settings.bother_klaptrap_model = bother_model_index
-    settings.beetle_model = racer_beetle
-    settings.rabbit_model = racer_rabbit
-    settings.panic_fairy_model = panic_fairy_model_index
-    settings.turtle_model = turtle_model_index
-    settings.panic_klaptrap_model = panic_klap_model_index
-    settings.seek_klaptrap_model = sseek_klap_model_index
-    settings.fungi_tomato_model = fungi_tomato_model_index
-    settings.caves_tomato_model = caves_tomato_model_index
-    settings.piano_burp_model = piano_burper
-    settings.spotlight_fish_model = spotlight_fish_model_index
-    settings.candy_cutscene_model = candy_model_index
-    settings.funky_cutscene_model = funky_model_index
-    settings.boot_cutscene_model = boot_model_index
-    settings.wrinkly_rgb = [255, 255, 255]
-    # Compute swap bitfield
-    swap_bitfield |= 0x10 if settings.rabbit_model == Model.Beetle else 0
-    swap_bitfield |= 0x20 if settings.beetle_model == Model.Rabbit else 0
-    swap_bitfield |= 0x40 if settings.fungi_tomato_model == Model.IceTomato else 0
-    swap_bitfield |= 0x80 if settings.caves_tomato_model == Model.Tomato else 0
-    # Write Models
-    ROM_COPY.seek(sav + 0x1B5)
-    ROM_COPY.writeMultipleBytes(settings.panic_fairy_model + 1, 1)  # Still needed for end seq fairy swap
-    ROM_COPY.seek(sav + 0x1E2)
-    ROM_COPY.write(swap_bitfield)
+    writeKongColors(settings)
+    
     settings.jetman_color = [0xFF, 0xFF, 0xFF]
     if settings.misc_cosmetics and settings.override_cosmetics:
         ROM_COPY.seek(sav + 0x196)
@@ -358,185 +141,17 @@ def apply_cosmetic_colors(settings: Settings):
             value = random.randint(brightness_threshold, 0xFF)
             jetman_color[channel] = value
         settings.jetman_color = jetman_color.copy()
-        melon_sprite = random.choice(melon_random_sprites)
-    settings.minigame_melon_sprite = melon_sprite
-    color_palettes = []
-    color_obj = {}
-    colors_dict = {}
-    kong_settings = [
-        KongPaletteSetting(
-            "dk",
-            0,
-            [
-                KongPalette("fur", 3724, PaletteFillType.block),
-                KongPalette("tie", 0x177D, PaletteFillType.block),
-                KongPalette("tie", 0xE8D, PaletteFillType.patch),
-            ],
-        ),
-        KongPaletteSetting(
-            "diddy",
-            1,
-            [
-                KongPalette("clothes", 3686, PaletteFillType.block),
-            ],
-        ),
-        KongPaletteSetting(
-            "lanky",
-            2,
-            [
-                KongPalette("clothes", 3689, PaletteFillType.block),
-                KongPalette("clothes", 3734, PaletteFillType.patch),
-                KongPalette("fur", 0xE9A, PaletteFillType.block),
-                KongPalette("fur", 0xE94, PaletteFillType.block),
-            ],
-        ),
-        KongPaletteSetting(
-            "tiny",
-            3,
-            [
-                KongPalette("clothes", 6014, PaletteFillType.block),
-                KongPalette("hair", 0xE68, PaletteFillType.block),
-            ],
-        ),
-        KongPaletteSetting(
-            "chunky",
-            4,
-            [
-                KongPalette("main", 3769, PaletteFillType.checkered, "other"),
-                KongPalette("main", 3687, PaletteFillType.block),
-            ],
-        ),
-        KongPaletteSetting(
-            "rambi",
-            5,
-            [
-                KongPalette("skin", 3826, PaletteFillType.block),
-            ],
-        ),
-        KongPaletteSetting(
-            "enguarde",
-            6,
-            [
-                KongPalette("skin", 3847, PaletteFillType.block),
-            ],
-        ),
-    ]
-
-    KONG_ZONES = {
-        "DK": ["Fur", "Tie"],
-        "Diddy": ["Clothes"],
-        "Lanky": ["Clothes", "Fur"],
-        "Tiny": ["Clothes", "Hair"],
-        "Chunky": ["Main", "Other"],
-        "Rambi": ["Skin"],
-        "Enguarde": ["Skin"],
-    }
-
+    
     if js.document.getElementById("override_cosmetics").checked or True:
         writeTransition(settings)
         writeCustomPortal(settings)
         writeCustomPaintings(settings)
         # randomizePlants(ROM_COPY, settings)  # Not sure how much I like how this feels
-        if js.document.getElementById("random_colors").checked:
-            for kong in KONG_ZONES:
-                for zone in KONG_ZONES[kong]:
-                    settings.__setattr__(f"{kong.lower()}_{zone.lower()}_colors", CharacterColors.randomized)
-        else:
-            for kong in KONG_ZONES:
-                for zone in KONG_ZONES[kong]:
-                    settings.__setattr__(
-                        f"{kong.lower()}_{zone.lower()}_colors",
-                        CharacterColors[js.document.getElementById(f"{kong.lower()}_{zone.lower()}_colors").value],
-                    )
-                    settings.__setattr__(
-                        f"{kong.lower()}_{zone.lower()}_custom_color",
-                        js.document.getElementById(f"{kong.lower()}_{zone.lower()}_custom_color").value,
-                    )
         settings.gb_colors = CharacterColors[js.document.getElementById("gb_colors").value]
         settings.gb_custom_color = js.document.getElementById("gb_custom_color").value
     else:
-        if settings.random_colors:
-            for kong in KONG_ZONES:
-                for zone in KONG_ZONES[kong]:
-                    settings.__setattr__(f"{kong.lower()}_{zone.lower()}_colors", CharacterColors.randomized)
         settings.gb_colors = CharacterColors.randomized
-
-    colors_dict = {}
-    for kong in KONG_ZONES:
-        for zone in KONG_ZONES[kong]:
-            colors_dict[f"{kong.lower()}_{zone.lower()}_colors"] = settings.__getattribute__(f"{kong.lower()}_{zone.lower()}_colors")
-            colors_dict[f"{kong.lower()}_{zone.lower()}_custom_color"] = settings.__getattribute__(f"{kong.lower()}_{zone.lower()}_custom_color")
-    for kong in kong_settings:
-        if kong.kong_index == 4:
-            if settings.kong_model_chunky == KongModels.disco_chunky:
-                kong.palettes = [
-                    KongPalette("main", 3777, PaletteFillType.sparkle),
-                    KongPalette("other", 3778, PaletteFillType.sparkle),
-                ]
-        settings_values = [
-            settings.kong_model_dk,
-            settings.kong_model_diddy,
-            settings.kong_model_lanky,
-            settings.kong_model_tiny,
-            settings.kong_model_chunky,
-        ]
-        if kong.kong_index >= 0 and kong.kong_index < len(settings_values):
-            if settings_values[kong.kong_index] in model_texture_sections:
-                base_setting = kong.palettes[0].name
-                kong.palettes = [
-                    KongPalette(base_setting, krusha_texture_replacement[kong.kong_index][0], PaletteFillType.block),  # krusha_skin
-                    KongPalette(base_setting, krusha_texture_replacement[kong.kong_index][1], PaletteFillType.kong),  # krusha_indicator
-                ]
-        base_obj = {"kong": kong.kong, "zones": []}
-        zone_to_colors = {}
-        for palette in kong.palettes:
-            arr = [DEFAULT_COLOR]
-            if palette.fill_type == PaletteFillType.checkered:
-                arr = ["#FFFF00", "#00FF00"]
-            elif palette.fill_type == PaletteFillType.kong:
-                arr = [getKongColor(settings, kong.kong_index)]
-            zone_data = {
-                "zone": palette.name,
-                "image": palette.image,
-                "fill_type": palette.fill_type,
-                "colors": arr,
-            }
-            for index in range(len(arr)):
-                base_setting = f"{kong.kong}_{palette.name}_colors"
-                custom_setting = f"{kong.kong}_{palette.name}_custom_color"
-                if index == 1:  # IS THE CHECKERED PATTERN
-                    base_setting = f"{kong.kong}_{palette.alt_name}_colors"
-                    custom_setting = f"{kong.kong}_{palette.alt_name}_custom_color"
-                if (settings.override_cosmetics and colors_dict[base_setting] != CharacterColors.vanilla) or (palette.fill_type == PaletteFillType.kong):
-                    color = None
-                    # if this palette color is randomized, and isn't krusha's kong indicator:
-                    if colors_dict[base_setting] == CharacterColors.randomized and palette.fill_type != PaletteFillType.kong:
-                        if base_setting in zone_to_colors:
-                            color = zone_to_colors[base_setting]
-                        else:
-                            color = f"#{format(randint(0, 0xFFFFFF), '06x')}"
-                            zone_to_colors[base_setting] = color
-                    # if this palette color is not randomized (but might be a custom color) and isn't krusha's kong indicator:
-                    elif palette.fill_type != PaletteFillType.kong:
-                        color = colors_dict[custom_setting]
-                        if not color:
-                            color = DEFAULT_COLOR
-                    # if this is krusha's kong indicator:
-                    else:
-                        color = getKongColor(settings, kong.kong_index)
-                    if color is not None:
-                        zone_data["colors"][index] = color
-                        base_obj["zones"].append(zone_data)
-                        color_palettes.append(base_obj)
-                        color_obj[f"{kong.kong} {palette.name}"] = color
-    settings.colors = color_obj
-    if len(color_palettes) > 0:
-        # this is just to prune the duplicates that appear. someone should probably fix the root of the dupe issue tbh
-        new_color_palettes = []
-        for pal in color_palettes:
-            if pal not in new_color_palettes:
-                new_color_palettes.append(pal)
-        convertColors(new_color_palettes)
+    
     # GB Shine
     if settings.override_cosmetics and settings.gb_colors != CharacterColors.vanilla:
         channels = []
@@ -678,60 +293,6 @@ def maskImageGBSpin(im_f, color: tuple, image_index: int):
         px_0[point[0], point[1]] = px[point[0], point[1]]
     return masked_im
 
-def hueShiftColor(color: tuple, amount: int, head_ratio: int = None) -> tuple:
-    """Apply a hue shift to a color."""
-    # RGB -> HSV Conversion
-    red_ratio = color[0] / 255
-    green_ratio = color[1] / 255
-    blue_ratio = color[2] / 255
-    color_max = max(red_ratio, green_ratio, blue_ratio)
-    color_min = min(red_ratio, green_ratio, blue_ratio)
-    color_delta = color_max - color_min
-    hue = 0
-    if color_delta != 0:
-        if color_max == red_ratio:
-            hue = 60 * (((green_ratio - blue_ratio) / color_delta) % 6)
-        elif color_max == green_ratio:
-            hue = 60 * (((blue_ratio - red_ratio) / color_delta) + 2)
-        else:
-            hue = 60 * (((red_ratio - green_ratio) / color_delta) + 4)
-    sat = 0 if color_max == 0 else color_delta / color_max
-    val = color_max
-    # Adjust Hue
-    if head_ratio is not None and sat != 0:
-        amount = head_ratio / (sat * 100)
-    hue = (hue + amount) % 360
-    # HSV -> RGB Conversion
-    c = val * sat
-    x = c * (1 - abs(((hue / 60) % 2) - 1))
-    m = val - c
-    if hue < 60:
-        red_ratio = c
-        green_ratio = x
-        blue_ratio = 0
-    elif hue < 120:
-        red_ratio = x
-        green_ratio = c
-        blue_ratio = 0
-    elif hue < 180:
-        red_ratio = 0
-        green_ratio = c
-        blue_ratio = x
-    elif hue < 240:
-        red_ratio = 0
-        green_ratio = x
-        blue_ratio = c
-    elif hue < 300:
-        red_ratio = x
-        green_ratio = 0
-        blue_ratio = c
-    else:
-        red_ratio = c
-        green_ratio = 0
-        blue_ratio = x
-    return (int((red_ratio + m) * 255), int((green_ratio + m) * 255), int((blue_ratio + m) * 255))
-
-
 def maskImageWithOutline(im_f, base_index, min_y, colorblind_mode, type=""):
     """Apply RGB mask to image with an Outline in a different color."""
     w, h = im_f.size
@@ -743,12 +304,11 @@ def maskImageWithOutline(im_f, base_index, min_y, colorblind_mode, type=""):
         im_dupe = brightener.enhance(2)
     im_f.paste(im_dupe, (0, min_y), im_dupe)
     pix = im_f.load()
-    mask = getRGBFromHash(getKongItemColor(colorblind_mode, base_index))
+    mask = getKongItemColor(colorblind_mode, base_index, True)
     if base_index == 2 or (base_index == 0 and colorblind_mode == ColorblindMode.trit):  # lanky or (DK in tritanopia mode)
-        border_color = getKongItemColor(colorblind_mode, Kongs.chunky)
+        mask2 = getKongItemColor(colorblind_mode, Kongs.chunky, True)
     else:
-        border_color = getKongItemColor(colorblind_mode, Kongs.diddy)
-    mask2 = getRGBFromHash(border_color)
+        mask2 = getKongItemColor(colorblind_mode, Kongs.diddy, True)
     contrast = False
     if base_index == 0:
         contrast = True
@@ -1003,49 +563,7 @@ def applyKongModelSwaps(settings: Settings) -> None:
                 base_im.paste(orange_im, (dim_offset, dim_offset), orange_im)
                 writeColorImageToROM(base_im, 25, switch_faces[index], 32, 32, False, TextureFormat.RGBA5551)
 
-def changeModelTextures(settings: Settings, kong_index: int):
-    """Change the textures associated with a model."""
-    settings_values = [
-        settings.kong_model_dk,
-        settings.kong_model_diddy,
-        settings.kong_model_lanky,
-        settings.kong_model_tiny,
-        settings.kong_model_chunky,
-    ]
-    if kong_index < 0 or kong_index >= len(settings_values):
-        return
-    model = settings_values[kong_index]
-    if model not in model_texture_sections:
-        return
-    for x in range(2):
-        file = kong_index_mapping[kong_index][x]
-        if file is None:
-            continue
-        krusha_model_start = js.pointer_addresses[5]["entries"][file]["pointing_to"]
-        krusha_model_finish = js.pointer_addresses[5]["entries"][file + 1]["pointing_to"]
-        krusha_model_size = krusha_model_finish - krusha_model_start
-        ROM_COPY = LocalROM()
-        ROM_COPY.seek(krusha_model_start)
-        indicator = int.from_bytes(ROM_COPY.readBytes(2), "big")
-        ROM_COPY.seek(krusha_model_start)
-        data = ROM_COPY.readBytes(krusha_model_size)
-        if indicator == 0x1F8B:
-            data = zlib.decompress(data, (15 + 32))
-        num_data = []  # data, but represented as nums rather than b strings
-        for d in data:
-            num_data.append(d)
-        # Retexture for colors
-        for tex_idx in model_texture_sections[model]["skin"]:
-            for di, d in enumerate(int_to_list(krusha_texture_replacement[kong_index][0], 2)):  # Main
-                num_data[tex_idx + di] = d
-        for tex_idx in model_texture_sections[model]["kong"]:
-            for di, d in enumerate(int_to_list(krusha_texture_replacement[kong_index][1], 2)):  # Belt
-                num_data[tex_idx + di] = d
-        data = bytearray(num_data)  # convert num_data back to binary string
-        if indicator == 0x1F8B:
-            data = gzip.compress(data, compresslevel=9)
-        LocalROM().seek(krusha_model_start)
-        LocalROM().writeBytes(data)
+
 
 def darkenDPad():
     """Change the DPad cross texture for the DPad HUD."""
@@ -1073,716 +591,6 @@ def darkenDPad():
     px_data = gzip.compress(px_data, compresslevel=9)
     ROM().seek(js.pointer_addresses[14]["entries"][187]["pointing_to"])
     ROM().writeBytes(px_data)
-
-def getValueFromByteArray(ba: bytearray, offset: int, size: int) -> int:
-    """Get value from byte array given an offset and size."""
-    value = 0
-    for x in range(size):
-        local_value = ba[offset + x]
-        value <<= 8
-        value += local_value
-    return value
-
-
-def getEnemySwapColor(channel_min: int = 0, channel_max: int = 255, min_channel_variance: int = 0) -> int:
-    """Get an RGB color compatible with enemy swaps."""
-    channels = []
-    for _ in range(2):
-        channels.append(random.randint(channel_min, channel_max))
-    min_channel = min(channels[0], channels[1])
-    max_channel = max(channels[0], channels[1])
-    bounds = []
-    if (min_channel - channel_min) >= min_channel_variance:
-        bounds.append([channel_min, min_channel])
-    if (channel_max - max_channel) >= min_channel_variance:
-        bounds.append([max_channel, channel_max])
-    if (len(bounds) == 0) or ((max_channel - min_channel) >= min_channel_variance):
-        # Default to random number pick
-        channels.append(random.randint(channel_min, channel_max))
-    else:
-        selected_bound = random.choice(bounds)
-        channels.append(random.randint(selected_bound[0], selected_bound[1]))
-    random.shuffle(channels)
-    value = 0
-    for x in range(3):
-        value <<= 8
-        value += channels[x]
-    return value
-
-
-class EnemyColorSwap:
-    """Class to store information regarding an enemy color swap."""
-
-    def __init__(self, search_for: list, forced_color: int = None):
-        """Initialize with given parameters."""
-        self.search_for = search_for.copy()
-        total_channels = [0] * 3
-        for color in self.search_for:
-            for channel in range(3):
-                shift = 8 * (2 - channel)
-                value = (color >> shift) & 0xFF
-                total_channels[channel] += value
-        average_channels = [int(x / len(self.search_for)) for x in total_channels]
-        self.average_color = 0
-        for x in average_channels:
-            self.average_color <<= 8
-            self.average_color += x
-        self.replace_with = forced_color
-        if forced_color is None:
-            self.replace_with = getEnemySwapColor(80, min_channel_variance=80)
-
-    def getOutputColor(self, color: int):
-        """Get output color based on randomization."""
-        if color not in self.search_for:
-            return color
-        if color == self.search_for[0]:
-            return self.replace_with
-        new_color = 0
-        total_boost = 0
-        for x in range(3):
-            shift = 8 * (2 - x)
-            provided_channel = (color >> shift) & 0xFF
-            primary_channel = (self.search_for[0] >> shift) & 0xFF
-            boost = 1  # Failsafe for div by 0
-            if primary_channel != 0:
-                boost = provided_channel / primary_channel
-            total_boost += boost  # Used to get an average
-        for x in range(3):
-            shift = 8 * (2 - x)
-            replacement_channel = (self.replace_with >> shift) & 0xFF
-            replacement_channel = int(replacement_channel * (total_boost / 3))
-            if replacement_channel > 255:
-                replacement_channel = 255
-            elif replacement_channel < 0:
-                replacement_channel = 0
-            new_color <<= 8
-            new_color += replacement_channel
-        return new_color
-
-
-def convertColorIntToTuple(color: int) -> tuple:
-    """Convert color stored as 3-byte int to tuple."""
-    return ((color >> 16) & 0xFF, (color >> 8) & 0xFF, color & 0xFF)
-
-
-def getLuma(color: tuple) -> float:
-    """Get the luma value of a color."""
-    return (0.299 * color[0]) + (0.587 * color[1]) + (0.114 * color[2])
-
-
-def adjustFungiMushVertexColor(shift: int):
-    """Adjust the special vertex coloring on Fungi Giant Mushroom."""
-    fungi_geo = bytearray(getRawFile(TableNames.MapGeometry, Maps.FungiForest, True))
-    DEFAULT_MUSHROOM_COLOR = (255, 90, 82)
-    NEW_MUSHROOM_COLOR = hueShiftColor(DEFAULT_MUSHROOM_COLOR, shift)
-    for x in range(0x27DA, 0x2839):
-        start = 0x25140 + (x * 0x10) + 0xC
-        channels = []
-        is_zero = True
-        for y in range(3):
-            val = fungi_geo[start + y]
-            if val != 0:
-                is_zero = False
-            channels.append(val)
-        if is_zero:
-            continue
-        visual_color = [int((x / 255) * DEFAULT_MUSHROOM_COLOR[xi]) for xi, x in enumerate(channels)]
-        luma = int(getLuma(visual_color))
-        # Diversify shading
-        luma -= 128
-        luma = int(luma * 1.2)
-        luma += 128
-        # Brighten
-        luma += 60
-        # Clamp
-        if luma < 0:
-            luma = 0
-        elif luma > 255:
-            luma = 255
-        # Apply shading
-        for y in range(3):
-            fungi_geo[start + y] = luma
-        fungi_geo[start + 3] = 0xFF
-    file_data = gzip.compress(fungi_geo, compresslevel=9)
-    ROM().seek(js.pointer_addresses[TableNames.MapGeometry]["entries"][Maps.FungiForest]["pointing_to"])
-    ROM().writeBytes(file_data)
-
-
-def writeMiscCosmeticChanges(settings):
-    """Write miscellaneous changes to the cosmetic colors."""
-    if settings.override_cosmetics:
-        enemy_setting = RandomModels[js.document.getElementById("random_enemy_colors").value]
-    else:
-        enemy_setting = settings.random_enemy_colors
-    if settings.misc_cosmetics:
-        # Melon HUD
-        data = {
-            7: [[0x13C, 0x147]],
-            14: [[0x5A, 0x5D]],
-            25: [
-                [getBonusSkinOffset(ExtraTextures.MelonSurface), getBonusSkinOffset(ExtraTextures.MelonSurface)],
-                [0x144B, 0x1452],
-            ],
-        }
-        shift = getRandomHueShift()
-        for table in data:
-            table_data = data[table]
-            for set in table_data:
-                for img in range(set[0], set[1] + 1):
-                    if table == 25:
-                        dims = (32, 32)
-                    else:
-                        dims = (48, 42)
-                    melon_im = getImageFile(table, img, table != 7, dims[0], dims[1], TextureFormat.RGBA5551)
-                    melon_im = hueShift(melon_im, shift)
-                    melon_px = melon_im.load()
-                    bytes_array = []
-                    for y in range(dims[1]):
-                        for x in range(dims[0]):
-                            pix_data = list(melon_px[x, y])
-                            red = int((pix_data[0] >> 3) << 11)
-                            green = int((pix_data[1] >> 3) << 6)
-                            blue = int((pix_data[2] >> 3) << 1)
-                            alpha = int(pix_data[3] != 0)
-                            value = red | green | blue | alpha
-                            bytes_array.extend([(value >> 8) & 0xFF, value & 0xFF])
-                    px_data = bytearray(bytes_array)
-                    if table != 7:
-                        px_data = gzip.compress(px_data, compresslevel=9)
-                    ROM().seek(js.pointer_addresses[table]["entries"][img]["pointing_to"])
-                    ROM().writeBytes(px_data)
-
-        # Shockwave Particles
-        shockwave_shift = getRandomHueShift()
-        for img_index in range(0x174F, 0x1757):
-            hueShiftImageContainer(25, img_index, 16, 16, TextureFormat.RGBA32, shockwave_shift)
-        if settings.colorblind_mode == ColorblindMode.off:
-            # Fire-based sprites
-            fire_shift = getRandomHueShift()
-            fires = (
-                [0x1539, 0x1553, 32],  # Fireball. RGBA32 32x32
-                [0x14B6, 0x14F5, 32],  # Fireball. RGBA32 32x32
-                [0x1554, 0x155B, 16],  # Small Fireball. RGBA32 16x16
-                [0x1654, 0x1683, 32],  # Fire Wall. RGBA32 32x32
-                [0x1495, 0x14A0, 32],  # Small Explosion, RGBA32 32x32
-                [0x13B9, 0x13C3, 32],  # Small Explosion, RGBA32 32x32
-            )
-            for sprite_data in fires:
-                for img_index in range(sprite_data[0], sprite_data[1] + 1):
-                    dim = sprite_data[2]
-                    hueShiftImageContainer(25, img_index, dim, dim, TextureFormat.RGBA32, fire_shift)
-            for img_index in range(0x29, 0x32 + 1):
-                hueShiftImageContainer(7, img_index, 32, 32, TextureFormat.RGBA32, fire_shift)
-            for img_index in range(0x250, 0x26F + 1):
-                hueShiftImageContainer(7, img_index, 32, 32, TextureFormat.RGBA32, fire_shift)
-            for img_index in range(0xA0, 0xA7 + 1):
-                hueShiftImageContainer(7, img_index, 32, 32, TextureFormat.RGBA5551, fire_shift)
-            # Blue Fire
-            for img_index in range(129, 138 + 1):
-                hueShiftImageContainer(7, img_index, 32, 32, TextureFormat.RGBA32, fire_shift)
-            # Number Game Numbers
-            COLOR_COUNT = 2  # 2 or 16
-            colors = [getRandomHueShift() for x in range(16)]
-            # vanilla_green = [2, 4, 5, 7, 9, 10, 12, 13]
-            vanilla_blue = [1, 3, 6, 8, 11, 14, 15, 16]
-            for x in range(16):
-                number_hue_shift = colors[0]
-                if COLOR_COUNT == 2:
-                    if (x + 1) in vanilla_blue:
-                        number_hue_shift = colors[1]
-                else:
-                    number_hue_shift = colors[x]
-                for sub_img in range(2):
-                    img_index = 0x1FE + (2 * x) + sub_img
-                    hueShiftImageContainer(7, img_index, 32, 32, TextureFormat.RGBA5551, number_hue_shift)
-            if COLOR_COUNT == 2:
-                hueShiftImageContainer(25, 0xC2D, 32, 32, TextureFormat.RGBA5551, colors[1])
-                hueShiftImageContainer(25, 0xC2E, 32, 32, TextureFormat.RGBA5551, colors[0])
-        boulder_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0x12F4, 1, 1372, TextureFormat.RGBA5551, boulder_shift)
-        for img_index in range(2):
-            hueShiftImageContainer(25, 0xDE1 + img_index, 32, 64, TextureFormat.RGBA5551, boulder_shift)
-
-    if enemy_setting != RandomModels.off:
-        # Barrel Enemy Skins - Random
-        klobber_shift = getRandomHueShift(0, 300)
-        kaboom_shift = getRandomHueShift()
-        for img_index in range(3):
-            px_count = 1404 if img_index < 2 else 1372
-            hueShiftImageContainer(25, 0xF12 + img_index, 1, px_count, TextureFormat.RGBA5551, klobber_shift)
-            hueShiftImageContainer(25, 0xF22 + img_index, 1, px_count, TextureFormat.RGBA5551, kaboom_shift)
-            if img_index < 2:
-                hueShiftImageContainer(25, 0xF2B + img_index, 1, px_count, TextureFormat.RGBA5551, kaboom_shift)
-        # Klump
-        klump_jacket_shift = getRandomHueShift()
-        klump_hatammo_shift = getRandomHueShift()
-        jacket_images = [
-            {"image": 0x104D, "px": 1372},
-            {"image": 0x1058, "px": 1372},
-            {"image": 0x1059, "px": 176},
-        ]
-        hatammo_images = [
-            {"image": 0x104E, "px": 1372},
-            {"image": 0x104F, "px": 1372},
-            {"image": 0x1050, "px": 1372},
-            {"image": 0x1051, "px": 700},
-            {"image": 0x1052, "px": 348},
-            {"image": 0x1053, "px": 348},
-        ]
-        for img_data in jacket_images:
-            hueShiftImageContainer(25, img_data["image"], 1, img_data["px"], TextureFormat.RGBA5551, klump_jacket_shift)
-        for img_data in hatammo_images:
-            hueShiftImageContainer(25, img_data["image"], 1, img_data["px"], TextureFormat.RGBA5551, klump_hatammo_shift)
-        # Kosha
-        kosha_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0x1232, 1, 348, TextureFormat.RGBA5551, kosha_shift)
-        hueShiftImageContainer(25, 0x1235, 1, 348, TextureFormat.RGBA5551, kosha_shift)
-        if enemy_setting == RandomModels.extreme:
-            kosha_helmet_int = getEnemySwapColor(80, min_channel_variance=80)
-            kosha_helmet_list = [
-                (kosha_helmet_int >> 16) & 0xFF,
-                (kosha_helmet_int >> 8) & 0xFF,
-                kosha_helmet_int & 0xFF,
-            ]
-            kosha_club_int = getEnemySwapColor(80, min_channel_variance=80)
-            kosha_club_list = [(kosha_club_int >> 16) & 0xFF, (kosha_club_int >> 8) & 0xFF, kosha_club_int & 0xFF]
-            for img in range(0x122E, 0x1230):
-                kosha_im = getImageFile(25, img, True, 1, 1372, TextureFormat.RGBA5551)
-                kosha_im = maskImageWithColor(kosha_im, tuple(kosha_helmet_list))
-                writeColorImageToROM(kosha_im, 25, img, 1, 1372, False, TextureFormat.RGBA5551)
-            for img in range(0x1229, 0x122C):
-                kosha_im = getImageFile(25, img, True, 1, 1372, TextureFormat.RGBA5551)
-                kosha_im = maskImageWithColor(kosha_im, tuple(kosha_club_list))
-                writeColorImageToROM(kosha_im, 25, img, 1, 1372, False, TextureFormat.RGBA5551)
-            if settings.colorblind_mode == ColorblindMode.off:
-                # Kremling
-                kremling_dimensions = [
-                    [32, 64],  # FCE
-                    [64, 24],  # FCF
-                    [1, 1372],  # fd0
-                    [32, 32],  # fd1
-                    [24, 8],  # fd2
-                    [24, 8],  # fd3
-                    [24, 8],  # fd4
-                    [24, 24],  # fd5
-                    [32, 32],  # fd6
-                    [32, 64],  # fd7
-                    [32, 64],  # fd8
-                    [36, 16],  # fd9
-                    [20, 28],  # fda
-                    [32, 32],  # fdb
-                    [32, 32],  # fdc
-                    [12, 28],  # fdd
-                    [64, 24],  # fde
-                    [32, 32],  # fdf
-                ]
-                while True:
-                    kremling_shift = getRandomHueShift()
-                    # Block red coloring
-                    if kremling_shift > 290:
-                        break
-                    if kremling_shift > -70 and kremling_shift < 228:
-                        break
-                    if kremling_shift < -132:
-                        break
-                for dim_index, dims in enumerate(kremling_dimensions):
-                    if dims is not None:
-                        hueShiftImageContainer(25, 0xFCE + dim_index, dims[0], dims[1], TextureFormat.RGBA5551, kremling_shift)
-            # Rabbit
-            rabbit_dimensions = [
-                [1, 1372],  # 111A
-                [1, 1372],  # 111B
-                [1, 700],  # 111C
-                [1, 700],  # 111D
-                [1, 1372],  # 111E
-                [1, 1372],  # 111F
-                [1, 1372],  # 1120
-                [1, 1404],  # 1121
-                [1, 348],  # 1122
-                [32, 64],  # 1123
-                [1, 688],  # 1124
-                [64, 32],  # 1125
-            ]
-            rabbit_shift = getRandomHueShift()
-            for dim_index, dims in enumerate(rabbit_dimensions):
-                if dims is not None:
-                    hueShiftImageContainer(25, 0x111A + dim_index, dims[0], dims[1], TextureFormat.RGBA5551, rabbit_shift)
-            # Snake
-            snake_shift = getRandomHueShift()
-            for x in range(2):
-                hueShiftImageContainer(25, 0xEF7 + x, 32, 32, TextureFormat.RGBA5551, snake_shift)
-        # Headphones Sprite
-        headphones_shift = getRandomHueShift()
-        for x in range(8):
-            hueShiftImageContainer(7, 0x3D3 + x, 40, 40, TextureFormat.RGBA5551, headphones_shift)
-        # Instruments
-        trombone_sax_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0xEA2, 32, 32, TextureFormat.RGBA5551, trombone_sax_shift)  # Shine
-        hueShiftImageContainer(25, 0x15AF, 40, 40, TextureFormat.RGBA5551, trombone_sax_shift)  # Trombone Icon
-        hueShiftImageContainer(25, 0x15AD, 40, 40, TextureFormat.RGBA5551, trombone_sax_shift)  # Sax Icon
-        hueShiftImageContainer(25, 0xBCC, 32, 64, TextureFormat.RGBA5551, trombone_sax_shift)  # Sax (Pad)
-        hueShiftImageContainer(25, 0xBCD, 32, 64, TextureFormat.RGBA5551, trombone_sax_shift)  # Sax (Pad)
-        hueShiftImageContainer(25, 0xBD0, 32, 64, TextureFormat.RGBA5551, trombone_sax_shift)  # Trombone (Pad)
-        hueShiftImageContainer(25, 0xBD1, 32, 64, TextureFormat.RGBA5551, trombone_sax_shift)  # Trombone (Pad)
-        triangle_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0xEBF, 32, 32, TextureFormat.RGBA5551, triangle_shift)  # Shine
-        hueShiftImageContainer(25, 0x15AE, 40, 40, TextureFormat.RGBA5551, triangle_shift)  # Triangle Icon
-        hueShiftImageContainer(25, 0xBCE, 32, 64, TextureFormat.RGBA5551, triangle_shift)  # Triangle (Pad)
-        hueShiftImageContainer(25, 0xBCF, 32, 64, TextureFormat.RGBA5551, triangle_shift)  # Triangle (Pad)
-        bongo_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0x1317, 1, 1372, TextureFormat.RGBA5551, bongo_shift)  # Skin
-        hueShiftImageContainer(25, 0x1318, 1, 1404, TextureFormat.RGBA5551, bongo_shift)  # Side
-        hueShiftImageContainer(25, 0x1319, 1, 1404, TextureFormat.RGBA5551, bongo_shift)  # Side 2
-        hueShiftImageContainer(25, 0x15AC, 40, 40, TextureFormat.RGBA5551, bongo_shift)  # Bongo Icon
-        hueShiftImageContainer(25, 0xBC8, 32, 64, TextureFormat.RGBA5551, bongo_shift)  # Bongo (Pad)
-        hueShiftImageContainer(25, 0xBC9, 32, 64, TextureFormat.RGBA5551, bongo_shift)  # Bongo (Pad)
-        if enemy_setting == RandomModels.extreme:
-            # Beanstalk
-            beanstalk_unc_size = [
-                0x480,
-                0x480,
-                0x480,
-                0x2B8,
-                0xAC0,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAF8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAF8,
-                0x578,
-                0xAB8,
-                0x578,
-                0x5F8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0xAB8,
-                0x578,
-                0xAB8,
-                0xAF8,
-                0xAB8,
-                0xAB8,
-                0x560,
-                0xAB8,
-                0x2B8,
-            ]
-            beanstalk_shift = getRandomHueShift()
-            for index, size in enumerate(beanstalk_unc_size):
-                hueShiftImageContainer(25, 0x1126 + index, 1, int(size >> 1), TextureFormat.RGBA5551, beanstalk_shift)
-        # Fairy Particles Sprites
-        fairy_particles_shift = getRandomHueShift()
-        for x in range(0xB):
-            hueShiftImageContainer(25, 0x138D + x, 32, 32, TextureFormat.RGBA32, fairy_particles_shift)
-        race_coin_shift = getRandomHueShift()
-        for x in range(8):
-            hueShiftImageContainer(7, 0x1F0 + x, 48, 42, TextureFormat.RGBA5551, race_coin_shift)
-        scoff_shift = getRandomHueShift()
-        troff_shift = getRandomHueShift()
-        scoff_data = {
-            0xFB8: 0x55C,
-            0xFB9: 0x800,
-            0xFBA: 0x40,
-            0xFBB: 0x800,
-            0xFBC: 0x240,
-            0xFBD: 0x480,
-            0xFBE: 0x80,
-            0xFBF: 0x800,
-            0xFC0: 0x200,
-            0xFC1: 0x240,
-            0xFC2: 0x100,
-            0xFB2: 0x240,
-            0xFB3: 0x800,
-            0xFB4: 0x800,
-            0xFB5: 0x200,
-            0xFB6: 0x200,
-            0xFB7: 0x200,
-        }
-        troff_data = {
-            0xF78: 0x800,
-            0xF79: 0x800,
-            0xF7A: 0x800,
-            0xF7B: 0x800,
-            0xF7C: 0x800,
-            0xF7D: 0x400,
-            0xF7E: 0x600,
-            0xF7F: 0x400,
-            0xF80: 0x800,
-            0xF81: 0x600,
-            0xF82: 0x400,
-            0xF83: 0x400,
-            0xF84: 0x800,
-            0xF85: 0x800,
-            0xF86: 0x280,
-            0xF87: 0x180,
-            0xF88: 0x800,
-            0xF89: 0x800,
-            0xF8A: 0x400,
-            0xF8B: 0x300,
-            0xF8C: 0x800,
-            0xF8D: 0x400,
-            0xF8E: 0x500,
-            0xF8F: 0x180,
-        }
-        for img in scoff_data:
-            hueShiftImageContainer(25, img, 1, scoff_data[img], TextureFormat.RGBA5551, scoff_shift)
-
-        # Scoff had too many bananas, and passed potassium poisoning onto Troff
-        # https://i.imgur.com/WFDLSzA.png
-        # for img in troff_data:
-        #     hueShiftImageContainer(25, img, 1, troff_data[img], TextureFormat.RGBA5551, troff_shift)
-        # Krobot
-        spinner_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0xFA9, 1, 1372, TextureFormat.RGBA5551, spinner_shift)
-        krobot_textures = [[[1, 1372], [0xFAF, 0xFAA, 0xFA8, 0xFAB, 0xFAD]], [[32, 32], [0xFAC, 0xFB1, 0xFAE, 0xFB0]]]
-        krobot_color_int = getEnemySwapColor(80, min_channel_variance=80)
-        krobot_color_list = [(krobot_color_int >> 16) & 0xFF, (krobot_color_int >> 8) & 0xFF, krobot_color_int & 0xFF]
-        for tex_set in krobot_textures:
-            for tex in tex_set[1]:
-                krobot_im = getImageFile(25, tex, True, tex_set[0][0], tex_set[0][1], TextureFormat.RGBA5551)
-                krobot_im = maskImageWithColor(krobot_im, tuple(krobot_color_list))
-                writeColorImageToROM(krobot_im, 25, tex, tex_set[0][0], tex_set[0][1], False, TextureFormat.RGBA5551)
-        # Jetman
-        for xi, x in enumerate(settings.jetman_color):
-            ROM().seek(settings.rom_data + 0x1E8 + xi)
-            ROM().writeMultipleBytes(x, 1)
-        # Blast Barrels
-        blast_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0x127E, 1, 1372, TextureFormat.RGBA5551, blast_shift)
-        for x in range(4):
-            hueShiftImageContainer(25, 0x127F + x, 16, 64, TextureFormat.RGBA5551, blast_shift)
-        hueShiftImageContainer(25, getBonusSkinOffset(ExtraTextures.BlastTop), 1, 1372, TextureFormat.RGBA5551, blast_shift)
-        # K Rool
-        red_cs_im = Image.new(mode="RGBA", size=(32, 32), color=convertColorIntToTuple(getEnemySwapColor()))
-        shorts_im = Image.new(mode="RGBA", size=(32, 32), color=convertColorIntToTuple(getEnemySwapColor()))
-        glove_im = Image.new(mode="RGBA", size=(32, 32), color=convertColorIntToTuple(getEnemySwapColor()))
-        krool_data = {
-            0x1149: red_cs_im,
-            0x1261: shorts_im,
-            0xDA8: glove_im,
-        }
-        if enemy_setting == RandomModels.extreme:
-            skin_im = Image.new(mode="RGBA", size=(32, 32), color=convertColorIntToTuple(getEnemySwapColor(80, min_channel_variance=80)))
-            krool_data[0x114A] = skin_im
-            krool_data[0x114D] = skin_im
-        for index in krool_data:
-            writeColorImageToROM(krool_data[index], 25, index, 32, 32, False, TextureFormat.RGBA5551)
-        toe_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0x126E, 1, 1372, TextureFormat.RGBA5551, toe_shift)
-        hueShiftImageContainer(25, 0x126F, 1, 1372, TextureFormat.RGBA5551, toe_shift)
-        if enemy_setting == RandomModels.extreme:
-            gold_shift = getRandomHueShift()
-            hueShiftImageContainer(25, 0x1265, 32, 32, TextureFormat.RGBA5551, gold_shift)
-            hueShiftImageContainer(25, 0x1148, 32, 32, TextureFormat.RGBA5551, gold_shift)
-        # Ghost
-        ghost_shift = getRandomHueShift()
-        for img in range(0x119D, 0x11AF):
-            px_count = 1372
-            if img == 0x119E:
-                px_count = 176
-            elif img == 0x11AC:
-                px_count = 688
-            hueShiftImageContainer(25, img, 1, px_count, TextureFormat.RGBA5551, ghost_shift)
-        # Funky
-        funky_shift = getRandomHueShift()
-        hueShiftImageContainer(25, 0xECF, 1, 1372, TextureFormat.RGBA5551, funky_shift)
-        hueShiftImageContainer(25, 0xED6, 1, 1372, TextureFormat.RGBA5551, funky_shift)
-        hueShiftImageContainer(25, 0xEDF, 1, 1372, TextureFormat.RGBA5551, funky_shift)
-        # Zinger
-        zinger_shift = getRandomHueShift()
-        zinger_color = hueShiftColor((0xFF, 0xFF, 0x0A), zinger_shift)
-        zinger_color_int = (zinger_color[0] << 16) | (zinger_color[1] << 8) | (zinger_color[2])
-        hueShiftImageContainer(25, 0xF0A, 1, 1372, TextureFormat.RGBA5551, zinger_shift)
-        # Mechazinger, use zinger color
-        for img_index in (0x10A0, 0x10A2, 0x10A4, 0x10A5):
-            hueShiftImageContainer(25, img_index, 1, 1372, TextureFormat.RGBA5551, zinger_shift)
-        hueShiftImageContainer(25, 0x10A3, 32, 32, TextureFormat.RGBA32, zinger_shift)
-        # Rings/DK Star
-        ring_shift = getRandomHueShift()
-        for x in range(2):
-            hueShiftImageContainer(25, 0xE1C + x, 1, 344, TextureFormat.RGBA5551, ring_shift)
-            hueShiftImageContainer(25, 0xD38 + x, 64, 32, TextureFormat.RGBA5551, ring_shift)
-        hueShiftImageContainer(7, 0x2EB, 32, 32, TextureFormat.RGBA5551, ring_shift)
-        # Buoys
-        for x in range(2):
-            hueShiftImageContainer(25, 0x133A + x, 1, 1372, TextureFormat.RGBA5551, getRandomHueShift())
-        # Trap Bubble
-        hueShiftImageContainer(25, 0x134C, 32, 32, TextureFormat.RGBA5551, getRandomHueShift())
-        # Spider
-        spider_shift = getRandomHueShift()
-        spider_dims = {
-            0x110A: (32, 64),
-            0x110B: (32, 64),
-            0x110C: (32, 64),
-            0x110D: (64, 16),
-            0x110E: (32, 64),
-            0x110F: (32, 64),
-            0x1110: (32, 64),
-            0x1111: (32, 64),
-            0x1112: (32, 64),
-            0x1113: (16, 32),
-            0x1114: (32, 32),
-            0x1115: (32, 32),
-            0x1116: (32, 32),
-            0x1117: (64, 16),
-            0x1118: (64, 32),
-            0x1119: (64, 32),
-        }
-        for img_index in spider_dims:
-            hueShiftImageContainer(
-                25,
-                img_index,
-                spider_dims[img_index][0],
-                spider_dims[img_index][1],
-                TextureFormat.RGBA5551,
-                spider_shift,
-            )
-
-        if enemy_setting == RandomModels.extreme:
-            # Army Dillo
-            dillo_px_count = {
-                0x102D: 64 * 32,
-                0x103A: 16 * 16,
-                0x102A: 24 * 24,
-                0x102B: 24 * 24,
-                0x102C: 1372,
-                0x103D: 688,
-                0x103E: 688,
-            }
-            dillo_shift = getRandomHueShift()
-            for img, px_count in dillo_px_count.items():
-                hueShiftImageContainer(25, img, 1, px_count, TextureFormat.RGBA5551, dillo_shift)
-
-        # Mushrooms
-        mush_man_shift = getRandomHueShift()
-        for img_index in (0x11FC, 0x11FD, 0x11FE, 0x11FF, 0x1200, 0x1209, 0x120A, 0x120B):
-            hueShiftImageContainer(25, img_index, 1, 1372, TextureFormat.RGBA5551, mush_man_shift)
-        for img_index in (0x11F8, 0x1205):
-            hueShiftImageContainer(25, img_index, 1, 692, TextureFormat.RGBA5551, mush_man_shift)
-        for img_index in (0x67F, 0x680):
-            hueShiftImageContainer(25, img_index, 32, 64, TextureFormat.RGBA5551, mush_man_shift)
-        hueShiftImageContainer(25, 0x6F3, 4, 4, TextureFormat.RGBA5551, mush_man_shift)
-        adjustFungiMushVertexColor(mush_man_shift)
-
-        # Enemy Vertex Swaps
-        blue_beaver_color = getEnemySwapColor(80, min_channel_variance=80)
-        enemy_changes = {
-            Model.BeaverBlue_LowPoly: EnemyColorSwap([0xB2E5FF, 0x65CCFF, 0x00ABE8, 0x004E82, 0x008BD1, 0x001333, 0x1691CE], blue_beaver_color),  # Primary
-            Model.BeaverBlue: EnemyColorSwap([0xB2E5FF, 0x65CCFF, 0x00ABE8, 0x004E82, 0x008BD1, 0x001333, 0x1691CE], blue_beaver_color),  # Primary
-            Model.BeaverGold: EnemyColorSwap([0xFFE5B2, 0xFFCC65, 0xE8AB00, 0x824E00, 0xD18B00, 0x331300, 0xCE9116]),  # Primary
-            Model.Zinger: EnemyColorSwap([0xFFFF0A, 0xFF7F00], zinger_color_int),  # Legs
-            Model.RoboZinger: EnemyColorSwap([0xFFFF00, 0xFF5500], zinger_color_int),  # Legs
-            Model.Candy: EnemyColorSwap(
-                [
-                    0xFF96EB,
-                    0x572C58,
-                    0xB86CAA,
-                    0xEB4C91,
-                    0x8B2154,
-                    0xD13B80,
-                    0xFF77C1,
-                    0xFF599E,
-                    0x7F1E4C,
-                    0x61173A,
-                    0x902858,
-                    0xA42E64,
-                    0x791C49,
-                    0x67183E,
-                    0x9E255C,
-                    0xC12E74,
-                    0x572C58,
-                    0xFF96EB,
-                    0xB86CAA,
-                ]
-            ),
-            Model.Laser: EnemyColorSwap([0xF30000]),
-            Model.Kasplat: EnemyColorSwap([0x8FD8FF, 0x182A4F, 0x0B162C, 0x7A98D3, 0x3F6CC4, 0x8FD8FF, 0x284581]),
-            # Model.BananaFairy: EnemyColorSwap([0xFFD400, 0xFFAA00, 0xFCD200, 0xD68F00, 0xD77D0A, 0xe49800, 0xdf7f1f, 0xa26c00, 0xd6b200, 0xdf9f1f])
-        }
-        if enemy_setting == RandomModels.extreme:
-            enemy_changes[Model.Klump] = EnemyColorSwap([0xE66B78, 0x621738, 0x300F20, 0xD1426F, 0xA32859])
-            dogadon_color = getEnemySwapColor(80, 160, min_channel_variance=80)
-            enemy_changes[Model.Dogadon] = EnemyColorSwap(
-                [
-                    0xFF0000,
-                    0xFF7F00,
-                    0x450A1F,
-                    0xB05800,
-                    0xFF3200,
-                    0xFFD400,
-                    0x4F260D,
-                    0x600F00,
-                    0x6A1400,
-                    0xAA0000,
-                    0xDF3F1F,
-                    0xFF251F,
-                    0x8F4418,
-                    0x522900,
-                    0xDF9F1F,
-                    0x3B0606,
-                    0x91121E,
-                    0x700C0D,
-                    0xFF5900,
-                    0xFF7217,
-                    0xFF7425,
-                    0xFF470B,
-                    0xA82100,
-                    0x4A0D18,
-                    0x580E00,
-                    0x461309,
-                    0x4C1503,
-                    0x780D0E,
-                    0xFFA74A,
-                    0x7E120F,
-                    0x700000,
-                    0xB64D19,
-                    0x883A13,
-                    0xBD351A,
-                    0xD42900,
-                    0xFF2A00,
-                    0x921511,
-                    0x9C662D,
-                    0xDF5F1F,
-                    0x9B1112,
-                    0x461F0A,
-                    0x4B0808,
-                    0x500809,
-                    0xA42000,
-                    0x5F0B13,
-                    0xBF6A3F,
-                    0x602E10,
-                    0x971414,
-                    0x422C15,
-                    0xFC5800,
-                    0x5C0D0B,
-                ],
-                dogadon_color,
-            )
-        for enemy in enemy_changes:
-            file_data = bytearray(getRawFile(5, enemy, True))
-            vert_start = 0x28
-            file_head = getValueFromByteArray(file_data, 0, 4)
-            disp_list_end = (getValueFromByteArray(file_data, 4, 4) - file_head) + 0x28
-            vert_end = (getValueFromByteArray(file_data, disp_list_end, 4) - file_head) + 0x28
-            vert_count = int((vert_end - vert_start) / 0x10)
-            for vert in range(vert_count):
-                local_start = 0x28 + (0x10 * vert)
-                test_rgb = getValueFromByteArray(file_data, local_start + 0xC, 3)
-                new_rgb = enemy_changes[enemy].getOutputColor(test_rgb)
-                for x in range(3):
-                    shift = 8 * (2 - x)
-                    channel = (new_rgb >> shift) & 0xFF
-                    file_data[local_start + 0xC + x] = channel
-            file_data = gzip.compress(file_data, compresslevel=9)
-            ROM().seek(js.pointer_addresses[5]["entries"][enemy]["pointing_to"])
-            ROM().writeBytes(file_data)
 
 def applyHelmDoorCosmetics(settings: Settings) -> None:
     """Apply Helm Door Cosmetic Changes."""
