+++ randomizer/Patching/Cosmetics/TextRando.py
@@ -1,3 +1,4 @@
+"""All code associated with cosmetic tweaks to text."""
 import random
 from randomizer.Patching.Patcher import LocalROM
 from randomizer.Patching.Lib import writeText, grabText
